// src/pages/system-configurations/maintain-message-parameters/type.ts
export interface IMessageParameters {
  key: number;
  displayName: string;
  maxRecipients: string;
  predefinedPeriod: string;
  maxEmail: string;
  emailAddress: string;
  emailGroup: string;
  userRole: string;
  tenantName: string;
  userStatus: string;
  userStatusLabel: string;
  accountStatus: string;
  accountStatusLabel: string;
}